+++ memory_test.go
@@ -10,7 +10,7 @@ func TestMemory(t *testing.T) {
 
 	memory.WriteString(12, "你好 world.")
 
-	fmt.Printf("%s\n", memory.ReadString(12, len([]byte("Hello world."))))
+	fmt.Printf("%s\n", memory.ReadString(12, uint32(len([]byte("Hello world.")))))
 
 	memory.WriteWord(1, 12)
